Move heartbeat into generators folder
