+++ kybra/compiler/kybra_generate/src/generators/canister_methods/mod.rs
@@ -1,0 +1,1 @@
+pub mod heartbeat;
