+++ kybra/compiler/kybra_generate/src/generators/mod.rs
@@ -1,4 +1,5 @@
 pub mod async_result_handler;
+pub mod canister_methods;
 pub mod ic_object;
 pub mod kybra_serde;
 pub mod rng_seed;
